Conflicts:
	calit2/GreenLight/GreenLight.h
	general/PluginTest/PluginTest.h

+++ calit2/ArtifactVis/vvtokenizer.cpp
@@ -123,7 +123,7 @@ int vvTokenizer::readChar()
     else
     {
       // Read next block:
-      blockUsed = (int)fread(data, 1, BLOCK_SIZE, fp);
+      blockUsed = fread(data, 1, BLOCK_SIZE, fp);
       cur = 0;
     }
   }
